refactored end_points.js and related code
